chore(ReDialog): 具体实现放到src文件中
