+++ src/components/ReBarcode/index.ts
@@ -1,7 +1,7 @@
-import reBarcode from "./src/index.vue";
+import _ReBarcode from "./src/index.vue";
 import { withInstall } from "@pureadmin/utils";
 
 /** 条形码组件 */
-export const ReBarcode = withInstall(reBarcode);
+export const ReBarcode = withInstall(_ReBarcode);
 
 export default ReBarcode;
