+++ src/components/ReCropper/index.ts
@@ -3,5 +3,4 @@ import { withInstall } from "@pureadmin/utils";
 
 /** 图片裁剪组件 */
 export const ReCropper = withInstall(reCropper);
-
 export default ReCropper;
