+++ src/components/ReDialog/src/type.ts
